requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-apscheduler==0.7.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
psycopg2==2.9.10
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
waitress==3.0.2
